Rename DbResultSetDeserializer to DbResultSetDeserializerImpl
